requirements.txt
langchain
python-dotenv
streamlit
unstructured
tiktoken
faiss-cpu
openai